pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "markdown-hub"
version = "0.1.0"
description = "MCP server for document conversion with Markdown as the hub"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "henry-1981" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "mcp[cli]",
    "markdown-it-py>=3.0.0",
    "mdit_py_plugins>=0.4.0",
    "playwright>=1.40.0",
    "python-docx>=1.1.0",
    "docxcompose>=1.4.0",
    "markitdown[all]",
    "pymupdf",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
pdf = ["opendataloader-pdf"]
vision = ["anthropic"]
hwp = ["pywin32"]

[project.urls]
Homepage = "https://github.com/henry-1981/md-hub"
Repository = "https://github.com/henry-1981/md-hub"

[tool.hatch.build.targets.wheel]
packages = ["src/md_hub"]

[project.scripts]
md-hub = "md_hub.server:main"
